backend/requirements.txt
fastapi
uvicorn
motor
pydantic
ultralytics
google-generativeai
python-multipart
pandas
python-dotenv
opencv-python-headless
